requirements.txt
asyncio
logging
pandas
numpy
ccxt
python-telegram-bot
